refactor(ai-builder): Address medium priority code review issues
- Remove unused getNodeVersion export from load-nodes.ts
- Extract validatePairwiseInputs and displayLocalResults helpers to reduce
  function complexity, remove eslint-disable complexity comments
- Add comment explaining --max-examples sentinel value pattern
- Rename pairwise-evaluator.ts → pairwise-judge-chain.ts for clarity
  (it's the LLM chain that acts as a judge)
- Rename pairwise-ls-evaluator.ts → pairwise-metrics-builder.ts for clarity
  (it builds LangSmith-compatible metrics from judge results)
- Update all imports and documentation references
